refactor: include post list modules into article post slices
