+++ app/records/page.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image';
 
 import EmptyPng from '^/public/status/empty.png';
 import { PostListItemProps } from '^/src/entities/post-list-item/props';
-import ArcadeRecordPostList from '^/src/features/arcade-record-post-list';
-import { getArcadeRecordPostList } from '^/src/features/arcade-record-post-list/data';
+import ArcadeRecordPostList from '^/src/features/arcade-record-article/arcade-record-post-list';
+import { getArcadeRecordPostList } from '^/src/features/arcade-record-article/arcade-record-post-list/data';
 
 export default async function RecordListPage() {
   const data = await getArcadeRecordPostList();
