+++ src/features/arcade-record-article/arcade-record-post-list/index.tsx
